refactor: 몬스터 HP 감소, 플레이어 EXP 데이터형 변화 (int >> double)
전투 중, 결과에서 바가 증가./감소하는 모습을 보여주고 싶었습니다.
int형으로 하면 소숫점 부분이 날아가기에 이를 정확히 표현해주게 하기 위해 double형으로 변경하였습니다.

+++ TextRpg/Player/PlayerData.cs
@@ -12,7 +12,7 @@ namespace TextRpg
         public string PlayerName { get; set; }
         public string PlayerClass { get; set; }
         public int Level { get; set; }
-        public int Exp { get; set; }
+        public double Exp { get; set; }
         public int MaxExp { get; set; }
         public int Gold { get; set; }
         public int Hp { get; set; }
